Changes were made principally to
1) Make MTDownload more readable and maintainable by significant reduction in the use of global variable to communicate between tasks
2) Significantly improve encapsulation in MTDownload
3) Hopefully make future changes easier to implement and test
4) Make tivodecode and ccextractor part of the initial download process and reduce disk usage

Created two new classes - MTTask and MTTaskChain

MTTask is designed to take care of all local resources and processes for each task as much as possible.  Customization of a task is done through the block handlers which are run at startup, completion, and cleanup.  There is also a handler for customization of progress calculation.

MTTaskChain is designed to be the plumber, as much as possible - i.e. it connects the tasks and organizes their execution based on both their structure and hints that are provided during their construction (requiresStandardInput, requiresStandardOutput) and their organization in the taskArray.  MTTaskChain will create and connect pipes as needed and will also break chains into subchains where no connection pipe is possible.  A chain consists of an array of arrays of tasks in which each array in the array contains tasks that can/should be run in parallel.

cTiVo is now ignoring the Simulencode preference entirely.  The UI for this still needs to be updated everywhere.

[ writeData] and the URL connection supports communicating either through an NSData buffer (urlBuffer) or a file on disk (_bufferFilePath).  
This allows cTiVo to initially try to keep the dataflow off the disk, except for final products, where possible.  But, the ability to do this 
depends on the processor being able to keep up with the data flow from the TiVo which is often not the case due to either a slow processor, 
fast network connection, of different tasks competing for processor resources.  When the processor falls too far behind and the memory 
buffer will become too large cTiVo will fall back to using files on the disk as a data buffer.

----Additional changes-----
- reset baseFileName on rescheduling a completed download so that it doesn't overwrite the completed download
- Add xattr to completed files in anticipation of adding checkpoints
- Fixed typo in README