Add ffmpeg wrapper, disable Crashlytics in Debug builds
This commit adds an ffmpeg wrapper script, called ffmpeg_edl_ac3.sh.
mencoder -edl was resulting in output files without audio for some
users, see #173.  In addition, it may be desirable for some users to use
ffmpeg for other purposes while still being able to cut commercials.

Therefore, a wrapper script for ffmpeg that translates an edl into
ffmpeg commands to encode multiple segments which are then merged
together.  A filter would be another way to do this, but is less
efficient, see #183.

Automatic detectinon 5.1 has also been a requested feature, see #151,
#179.  This wrapper script checks for 5.1 audio in the input and adds an
ac3 stream in the output if found.

There is still a pending issue where cTiVo seems to kill the commercial
task if it thinks the encoding has stalled, #184.  Committing anyway
as I think we need the script there to be able to reproduce.

The first thing I noticed when I set about resolving #183 was that I was
unable to build and the installation instructions for Crashlytics
require a bulidable project.  So, I moved some build settings around in
a way that builds out of the box without Crashlytics and should still
compile and link to Crashlytics in Release builds.  *However,
Crashlytics still needs to be tested in a Release build to make sure the
build still works and it is still functional!*.  See #158

Sorry for rolling both of these into the same change...

cTiVo/MTAppDelegate.m
cTiVo/MTTask.m :
  - Wrap crashlytics code in #ifndef DEBUG

cTiVo.xcodeproj/project.pbxproj :
  - Remove "Crashlytics.sh" from "Copy Bundle Resources" build phase and added the copy to the run script build phase, only copies if $CONFIGURATION = Release
  - Only run Crashlytics.sh in the run script build phase if $CONFIGURATION = Release
  - Remove Crashlytics and Fabric frameworks from "Linked rameworks and Libraries" general build setting
  - Add "-framework Crashlytics -framework Fabric" to "Other Linker Flags" build setting under release only
  - Add cTiVoScripts to project
  - Add ffmpeg_edl_ac3.sh to Copy Executables build phase

cTiVo/formats.plist :
  - add new built-in format for the wrapper script (ffmpeg_edl_ac3.sh)

cTiVo/cTiVoScripts/ffmpeg_edl_ac3.sh :
  - new file and first cTiVoScript
  - attempts to effectively add an -edl option to ffmpeg
  - attempts to detect 5.1 in any of the input audio streams and outputs a 5.1 ac3 audio stream if found
  - tested on several OTA broadcast styles, multiple parallel encodings

Resolves: #183
See also: #151, #158, #173, #179, #184