Fix typos in wrapper script
The previous version of the wrapper script was largely a refactor of
previous versions in addition to fixing issues when the path to the
cTiVo bundle  contained spaces.  However, it was untested and turned out
to be non-functional, so this version fixes several typos.

I also added back the removal of the output file before proceeding with
the encode; otherwise ffmpeg prompts yes/no for removal of the output
file which unneccisarily complicates things when running standalone for
testing or other reasons.  cTiVo shouldn't be calling the wrapper script
with a pre-existing output file anyway so having teh rm -rf shouldn't
hurt anything.