Fix segment-out-of-order bug + minor enhancements
Oops, in renaming the segment file names to be based on the cut times
and dynamically generating the filename list for ffmpeg, I accidentally
broke segment ordering.

In addition to fixing the above issue, this change bases the tmpdir
_name_ on the output file while keeping the tmpdir _location_ based on
the input.  Also adding the printing of a message and exiting if no
video streams found, both requested by Hugh, see #186.

Moving some global settings to top-of-file,
$merge_start_percent and $no_video_stream_message.

Also changed $merge_start_percent from 90 to 95 which should smooth out
the progress bar a bit; 95 is even probably a little low as merge time
<< encode time.