Escape quotes, handle multiple streams correctly
Hugh found that the concat or "merge" step was failing to find the
segment files if the tmpdir (based on input file dirname and output file
basename) had single quotes in it.  Added in a search-and-replace to
escape single quotes in the absolute path of the segment files.
http://ffmpeg.org/ffmpeg-utils.html#Quoting-and-escaping

I realized that the concat ffmpeg call was mixing up streams if the
video and audio streams in the input file are not in the same order as
the map arguments.  In fact, ffmpeg should be called without the map or
ac3 opts for the concat step.  Simply adding -map 0 -c copy copies
everything as-is from the segments to the merged file.

As requested by Hugh in #183, the ffmpeg log file contents if the output
file is nto written.  We may want to print to stderr instead of stdout?

As discussed in #183, there are plenty of other commands inside the
ffmpeg wait loop besides sleep that increment the PID, so replacing
sleep with the read hack doesn't really buy us anything (the read
commmand wasn't working anyhow, it was just generating lots of errors)
So, put sleep back and instead slow the ffmpeg wait loop down, added
"$sleep_duration" configurable global to top of file, set to 10 for now.

I capped the progress monitor at max_percent because I was seeing it go
slightly above in some cases, maybe due to negative timestamps at the
start of the stream?

Some more minor changes:
  - Moved segments to their own subdir under $tmpdir.
  - Format the segment filenames by '%06d_%06d', ss, to, should result
    in proper alphabetical sorting for videos of reasonable length
  - Renamed ffmpeg_merge.txt to segment_list.txt
  - Bumped $merge_start_percent to 98 which seems like it should even
    out the two progress lines better.