Fix several missing items in subscription dict
In trying to figure out how cTiVo remembers which shows were downloaded
for a subscription, I came across MTSubscription.prevRecorded but
realized it was never getting saved.  Turns out this was because
kMTExportTextMetaData was never getting initialized so was always nil
for subscriptions.  Everything after it in the dictionary saved to
kMTSubscriptionList was getting lost, including exportSubtitles,
preferredTivo, HDOnly, SDOnly, and prevRecorded.

Simply initializing kMTExportTextMetaData in the user defaults fixes
this.